Add app creation portal and deletion protection
I've created a sub-app that functions as a user portal for creating
client apps to use the service for authentication. It's CRUD-complete
and, of course, uses itself for authentication.

Apps and User accounts have deletion protection that requires sending
an authentication code to the user's email.

+++ src/app/portal/crud/user_crud.py
@@ -3,7 +3,9 @@ from typing import Optional
 import mongox
 from fastapi import HTTPException
 
+from app.portal.crud import clientapp_crud
 from app.portal.models.user_model import User
+from app.portal.services import deletion_protection
 
 
 async def get_user_by_email(email: str) -> Optional[User]:
@@ -29,4 +31,31 @@ async def check_or_create_user_from_email(email: str) -> User:
 
 
 async def delete_user(user: User) -> None:
+    if user.deletion_protection:
+        raise HTTPException(
+            status_code=400,
+            detail="You need to turn off deletion protection to delete your account.",
+        )
+    await clientapp_crud.delete_user_apps(user)
     await user.delete()
+
+
+async def update_user(user: User, name: str) -> User:
+    user.name = name
+    await user.save()
+    return user
+
+
+async def disable_deletion_protection(user: User, dp_code: str) -> User:
+    """
+    Disable deletion protection for a user account.
+
+    :param user:  the user to disable deletion protection for
+    :param code:  the one time use code emailed to the user
+    :return:      the user object
+    """
+    if not deletion_protection.verify_dp_code(user, "account", dp_code):
+        raise HTTPException(status_code=400, detail="Invalid deletion protection code.")
+    user.deletion_protection = False
+    await user.save()
+    return user
